axis

+++ prepare_lr.py
@@ -10,8 +10,8 @@ import yaml
 
 def add_padding(arr, length):
     n1, n2, n3 = arr.shape
-    padding = np.ones((n1, n2, length))*np.mean(arr)
-    return np.concatenate(arr, padding, axis=2)
+    padding = np.ones((n1, n2, length), dtype=arr.dtype)*np.mean(arr)
+    return np.concatenate((arr, padding), axis=2)
 
 def find_2nd(string, substring):
     '''
